requirements.txt
geopy~=2.4.1
jupyter~=1.1.1
pip~=25.2
pathlib~=1.0.1
configparser~=7.2.0
attrs~=25.3.0
wheel~=0.38.4
filelock~=3.19.1
setuptools~=78.1.0
packaging~=25.0
tomli~=2.2.1
typing_extensions~=4.14.1
Cython~=3.1.3
Pygments~=2.19.2
platformdirs~=4.4.0
pillow~=11.3.0
typing_extensions~=4.14.1
piexif~=1.1.3